Finished moving all graph_handling.py funcs over to model_history.py; just need to debug and refactor the __str__ methods now, and make final choices about output. Also be consistent about whether modules vs. passes are what gets seen.
