+++ torchlens/user_funcs.py
@@ -3,7 +3,6 @@ from typing import List, Optional, Union
 import torch
 from torch import nn
 
-from torchlens.graph_handling import ModelHistory
 from torchlens.helper_funcs import warn_parallel
 from torchlens.model_funcs import get_all_submodules, run_model_and_save_specified_activations
 from torchlens.validate import validate_model_history
